requirements.txt
-e .
pytest
aws_cdk.core
boto3